# api/rest/openapi.yaml
openapi: 3.1.0
servers:
  - url: /v1
info:
  description: |-
    This is an API to talk to this SANSA Space Science - EDA instrumentation
    management service.
  version: 1.0.0
  title: SSS-EDA OpenAPI
  termsOfService: 'https://sansa.dev/terms/'
  contact:
    email: sschoeman94@gmail.com
  license:
    name: Apache 2.0
    url: 'http://sansa.dev/licenses/LICENSE.html'
...
paths:
  /instrument:
    post:
      tags:
        - instrument
      summary: Register a new instrument
      description: Register a new instrument
      operationId: addInstrument
      responses:
        '200':
          description: Successful operaton
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instrument'
        '405':
          description: Invalid input
      security:
        - auth:
          - 'write:instrument'
          - 'read:instrument'
      requestBody:
        description: Register a new instrument
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/Instrument'
...
components:
  schemas:
    Instrument:
      properties:
        id:
          type: string
        name:
          type: string
        siteID:
          type: string
        instrumentTypeID:
          type: string
...